move classes a bit
