+++ anvil-appcompat-v7/src/gen-appcompatV7-1.1.0/kotlin/dev/inkremental/dsl/androidx/appcompat/AppcompatV7Setter.kt
@@ -54,11 +54,11 @@ import kotlin.Unit
 /**
  * DSL for creating views and settings their attributes.
  * This file has been generated by
- * {@code gradle AppCompatv7Setter}
- * It contains views and their setters for the library appcompat-v7.
+ * {@code gradle AppcompatV7Setter}
+ * It contains views and their setters for the library appcompat-v7-1.1.0.
  * Please, don't edit it manually unless for debugging.
  */
-object AppCompatv7Setter : Inkremental.AttributeSetter<Any> {
+object AppcompatV7Setter : Inkremental.AttributeSetter<Any> {
   override fun set(
     v: View,
     name: String,
